+++ CHANGES.rst
@@ -81,6 +81,8 @@ Accessibility
 
 - Improve country input accessibility (:pr:`6551`, thanks :user:`foxbunny`)
 - Reimplement Checkbox to make it programmatically focusable (:pr:`6528`, thanks :user:`foxbunny`)
+- Implement a ``RadioButton`` component to replace the SUI radio button in order to improve
+  keyboard support (thanks :user:`foxbunny`)
 
 Internal Changes
 ^^^^^^^^^^^^^^^^
@@ -90,6 +92,8 @@ Internal Changes
 - Add additional signals related to videoconferences and their event links (:pr:`6475`)
 - Videoconference plugins now need to implement a ``delete_room`` method (:pr:`6475`)
 - Support translator comments when extracting translatable strings (:pr:`6620`)
+- ``renderAsFieldset`` option in the registration field registry can now be a function that
+  returns a boolean (thanks :user:`foxbunny`)
 
 
 Version 3.3.4